pyproject.toml
[tool.poetry]
name = "portfolio-app"
version = "0.1.0"
description = "This app is a template to create a resume or portfolio"
authors = ["Young Padawan Flavia Ferreira <flavsferr@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.30.0"
pandas = "^2.1.4"
glob2 = "^0.7"
pybase64 = "^1.3.2"
streamlit-elements = "^0.1.0"
streamlit-option-menu = "^0.3.6"
streamlit-lottie = "^0.0.5"
pyyaml = "^6.0.1"
ruamel-yaml = "^0.18.5"

[tool.poetry.scripts]
portfolio = "portfolio_app.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
